feat: support redisson compatible springboot integration
