+++ arex-instrumentation/redis/arex-redission-v316/pom.xml
@@ -10,27 +10,20 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>arex-redis-common</artifactId>
+    <artifactId>arex-redission-v316</artifactId>
 
     <dependencies>
-        <dependency>
-            <groupId>io.lettuce</groupId>
-            <artifactId>lettuce-core</artifactId>
-            <version>6.1.8.RELEASE</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>io.arex</groupId>
-            <artifactId>arex-common</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
         <dependency>
             <groupId>org.redisson</groupId>
             <artifactId>redisson</artifactId>
             <version>3.17.3</version>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>io.arex</groupId>
+            <artifactId>arex-redis-common</artifactId>
+            <version>${project.version}</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
-
 </project>
